# feat(core): Add support for modules in realms.