Part of #13239.