@andreubotella 2022-09-05
With these changes, `deno bench` now hangs forever after the benchmarks are run. This seems to be because the code that drives the benchmarks waits on a mpsc receiver until all senders have been dropped (https://github.com/denoland/deno/blob/16dbf4adc390c9fb7656372b42811c1929e755dd/cli/tools/bench.rs#L407), and one of them is being kept alive. This is because now a realm's module map is stored in the context slot (rather than in the isolate slot), and because of denoland/rusty_v8#1066 it ends up not getting dropped even after the isolate is dropped.