@andreubotella 2022-10-16
#16286 changed the module map, that was in a slot in the isolate backed by the `IsolateAnnex`, to use a raw V8 slot in the isolate. But this PR changes the module map to be a slot in the context's annex, so some of the perf gains from that PR will regress. But I expect #16215 to land before this PR, after which the module map could be stored in a raw slot in the context.