@andreubotella 2022-10-17
While starting to look into writing tests for this, I realized that in this PR I'm using a single `ModuleLoader` instance for all of an isolate's realms. There are separate module maps per realm, so in that sense they're separate, even if they share a loader, but since `ModuleLoader` implementations might have state, we might need to consider the implications. If we add an `Option<Fn() -> Rc<dyn ModuleLoader>>` field to `RuntimeOptions` (plus the current `module_loader` field for the main realm), we would also let embedders disallow module imports in `ShadowRealm`s, which might be worth considering.