@andreubotella 2023-01-17
PTAL @bartlomieju @littledivy. Although this is probably not ready for merging, and the comments https://github.com/denoland/deno/pull/15760#issuecomment-1279953186 and https://github.com/denoland/deno/pull/15760#issuecomment-1281613105 are still issues, it would be good to get some initial reviews.