@bartlomieju 2023-01-21
> PTAL @bartlomieju @littledivy. Although this is probably not ready for merging, and the comments [#15760 (comment)](https://github.com/denoland/deno/pull/15760#issuecomment-1279953186) and [#15760 (comment)](https://github.com/denoland/deno/pull/15760#issuecomment-1281613105) are still things to fix, it would be good to get some initial reviews.

Sorry Andreu, I was out for most of the week. I'll put it into my TODO list for the next week and provide initial review.