@andreubotella 2023-01-31
> The question about `ModuleLoader` is an interesting one and I don't have a good answer right now. Does spec say anything about it?

The HTML spec requires a level of "caching" in module fetches between a page's main and shadow realms, such that if a module is imported in the main realm it should not be re-fetched when imported from the shadow realm, and vice versa. I haven't looked at this code in a long time, but last time I checked Deno did some caching for module sources that could be directly reused here.

But as I was looking at the spec, I realized that import maps were recently merged into the HTML spec, and it's not clear if anyone has looked in any depth at how they interact with `ShadowRealm`. Merging the HTML PR as is would mean that import maps don't apply in `ShadowRealm`s, which seems reasonable, [but I asked in a comment just in case](https://github.com/whatwg/html/pull/5339#issuecomment-1409604285). In this case, shadow realms would need a different module loader from the main realm, since they'd have to have a flag to not use import maps.

So I guess we should probably change `RuntimeOptions` to have a module loader for the main realm, and one for all shadow realms. Since they're `Option<Rc<dyn ModuleLoader>>`, both can be references to the same module loader object.


> Also we need to figure out what we're gonna do with snapshotting. I think for the first pass we should not care about additional realms and just make sure that snapshotting still works with a "global realm"

WFM