@andreubotella 2023-02-05
Since `JsRuntime::create_realm` can be used to create "regular" (non-shadow) realms, for example if we ever support the Node.js `vm` module, I thought it was better to require the realm creator to provide a module loader. Then in #16211 (when I find time to rebase that PR) we can have a module loader that gets used for every shadow reallm.