@bartlomieju 2023-02-06
Let's wait on https://github.com/denoland/deno/pull/17648 before we merge this PR.