@andreubotella 2023-04-08
I think this PR is now ready for review. The commit message contains a somewhat detailed list of changes.

I thought changing the module map to be stored on a context slot would make things slower, but the basic `exec_time` benchmarks don't show a difference. @littledivy, are there benchmarks specific to module loading that I should be running instead?

-----

Main branch:

```
Benchmark 1: /home/abotella/Projects/forks/denoland/deno/target/release/deno run --reload cli/tests/testdata/run/002_hello.ts 
  Time (mean ± σ):      18.1 ms ±   0.9 ms    [User: 11.2 ms, System: 7.9 ms]
  Range (min … max):    16.9 ms …  23.3 ms    157 runs

Benchmark 2: /home/abotella/Projects/forks/denoland/deno/target/release/deno run --reload cli/tests/testdata/run/003_relative_import.ts
  Time (mean ± σ):      17.8 ms ±   0.6 ms    [User: 11.2 ms, System: 7.7 ms]
  Range (min … max):    17.1 ms …  21.0 ms    156 runs
 
Benchmark 3: /home/abotella/Projects/forks/denoland/deno/target/release/deno run cli/tests/testdata/run/002_hello.ts
  Time (mean ± σ):      18.0 ms ±   1.3 ms    [User: 11.1 ms, System: 8.0 ms]
  Range (min … max):    16.9 ms …  27.8 ms    156 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 4: /home/abotella/Projects/forks/denoland/deno/target/release/deno run cli/tests/testdata/run/003_relative_import.ts
  Time (mean ± σ):      18.1 ms ±   1.0 ms    [User: 11.1 ms, System: 8.1 ms]
  Range (min … max):    17.1 ms …  25.8 ms    159 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 5: /home/abotella/Projects/forks/denoland/deno/target/release/deno run cli/tests/testdata/run/error_001.ts ; test $? -eq 1
  Time (mean ± σ):      18.3 ms ±   0.7 ms    [User: 11.1 ms, System: 8.2 ms]
  Range (min … max):    17.3 ms …  20.7 ms    156 runs
 
Benchmark 6: /home/abotella/Projects/forks/denoland/deno/target/release/deno run --reload --no-check cli/tests/testdata/run/002_hello.ts 
  Time (mean ± σ):      18.7 ms ±   1.2 ms    [User: 11.7 ms, System: 8.1 ms]
  Range (min … max):    17.0 ms …  28.6 ms    152 runs
 
Benchmark 7: /home/abotella/Projects/forks/denoland/deno/target/release/deno run --allow-read cli/tests/testdata/workers/bench_startup.ts
  Time (mean ± σ):     732.8 ms ±  10.0 ms    [User: 600.6 ms, System: 182.6 ms]
  Range (min … max):   720.6 ms … 754.7 ms    10 runs

Benchmark 8: /home/abotella/Projects/forks/denoland/deno/target/release/deno run --allow-read cli/tests/testdata/workers/bench_round_robin.ts 
  Time (mean ± σ):     173.4 ms ±   8.8 ms    [User: 367.1 ms, System: 51.8 ms]
  Range (min … max):   146.8 ms … 183.6 ms    16 runs
```

This PR:

```
Benchmark 1: /home/abotella/Projects/forks/denoland/deno/target/release/deno run --reload cli/tests/testdata/run/002_hello.ts 
  Time (mean ± σ):      18.5 ms ±   2.2 ms    [User: 11.4 ms, System: 8.2 ms]
  Range (min … max):    16.8 ms …  34.5 ms    155 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Benchmark 2: /home/abotella/Projects/forks/denoland/deno/target/release/deno run --reload cli/tests/testdata/run/003_relative_import.ts 
  Time (mean ± σ):      18.1 ms ±   0.6 ms    [User: 11.0 ms, System: 8.2 ms]
  Range (min … max):    17.1 ms …  20.7 ms    159 runs
 
Benchmark 3: /home/abotella/Projects/forks/denoland/deno/target/release/deno run cli/tests/testdata/run/002_hello.ts 
  Time (mean ± σ):      17.9 ms ±   0.8 ms    [User: 11.1 ms, System: 7.9 ms]
  Range (min … max):    16.9 ms …  24.4 ms    159 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 4: /home/abotella/Projects/forks/denoland/deno/target/release/deno run cli/tests/testdata/run/003_relative_import.ts 
  Time (mean ± σ):      18.2 ms ±   1.0 ms    [User: 11.0 ms, System: 8.3 ms]
  Range (min … max):    17.1 ms …  23.1 ms    160 runs
 
Benchmark 5: /home/abotella/Projects/forks/denoland/deno/target/release/deno run cli/tests/testdata/run/error_001.ts ; test $? -eq 1
  Time (mean ± σ):      18.3 ms ±   0.8 ms    [User: 10.9 ms, System: 8.4 ms]
  Range (min … max):    17.3 ms …  24.8 ms    149 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 6: /home/abotella/Projects/forks/denoland/deno/target/release/deno run --reload --no-check cli/tests/testdata/run/002_hello.ts 
  Time (mean ± σ):      17.9 ms ±   0.5 ms    [User: 11.1 ms, System: 7.9 ms]
  Range (min … max):    17.2 ms …  20.2 ms    153 runs
 
Benchmark 7: /home/abotella/Projects/forks/denoland/deno/target/release/deno run --allow-read cli/tests/testdata/workers/bench_startup.ts 
  Time (mean ± σ):     731.0 ms ±   5.8 ms    [User: 604.0 ms, System: 179.8 ms]
  Range (min … max):   719.9 ms … 739.5 ms    10 runs
 
Benchmark 8: /home/abotella/Projects/forks/denoland/deno/target/release/deno run --allow-read cli/tests/testdata/workers/bench_round_robin.ts 
  Time (mean ± σ):     174.2 ms ±   5.7 ms    [User: 375.9 ms, System: 53.5 ms]
  Range (min … max):   162.1 ms … 185.3 ms    16 runs
```